Производит модуляцию файлов
